add / remove in bsddb views
when the db or a result set is modified with add or remove, the view is updated
accordingly. The view BTree now contains keys formatted like as

    <collation_value>\O<key>  -> <key>

This makes it trivial to disambiguate view entries, and more efficient to
search for a specific entry; in addition, it removes the need for an external
repository of counters.

imported from: gobry@pybliographer.org--public/pybliographer--devel--1.3--patch-71