tagged reader as a state machine
Attempt to write the tagged reader as a state machine, where the person wishing
to implement a new parser for a format will only need to handle simple cases on
a per-line basis. This is a bit more verbose than the previous method, but
looks more powerful too.

imported from: gobry@pybliographer.org--public/pybliographer--devel--1.3--patch-75