method to get txo items by name
It is now possible to use a method txogroup.byname (name) to fetch txo items.
The name is defined in the txo item as being the name in language 'C'. This
language is officially promoted as the language used by (C)ode, not by people.

imported from: gobry@pybliographer.org--public/pybliographer--devel--1.3--patch-168