parse weird pubmed entries
In some cases, pubmed will return records with no journal title in the Journal field,
but a title is still available in the MedlineTA field. The code now falls back to this
version if it is the only one available.

Added a convenience method to check for incomplete Attributes.

imported from: gobry@pybliographer.org--2007/pyblio-core--devel--1.3--patch-15